+++ demo/graphviz_basic/graphviz_demo.py
@@ -18,7 +18,7 @@ from pathlib import Path
 
 from onya.graph import graph
 from onya.serial import graphviz
-from onya.serial.literate_lex import LiterateParser
+from onya.serial.literate import LiterateParser
 
 
 def demo_basic():